#680 - SkyAtmosphere 리팩토링: Aerial Perspective 함수 제거 및 AtmosphereParameters를 SkyAtmosphere 구조체로 대체, 셰이더 바인딩 최적화 (V4.1.0-Alpha-rnd)

+++ src/display/skyAtmosphere/core/skyAtmosphereFn.wgsl
@@ -8,7 +8,7 @@
 const MAX_TAU: f32 = 50.0;
 
 // [KO] 레이-구체 교차점 계산
-fn get_ray_sphere_intersection(ray_origin: vec3<u32>, ray_dir: vec3<f32>, sphere_radius: f32) -> f32 {
+fn get_ray_sphere_intersection(ray_origin: vec3<f32>, ray_dir: vec3<f32>, sphere_radius: f32) -> f32 {
     let b = dot(ray_origin, ray_dir);
     let c = dot(ray_origin, ray_origin) - sphere_radius * sphere_radius;
     let delta = b * b - c;
